+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.5.12
+
+- Bug fix [[BUG] Ignore Item Weight and Inherent Weight does not work](https://github.com/p4535992/inventory-plus/issues/10)
+
 ### 0.5.11
 
 - Bug fix [[BUG] Conflict with party Inventory](https://github.com/p4535992/inventory-plus/issues/9)